Additional getToday functionality
The stock market is closed on weekends so when the API tries to fetch on a Saturday or Sunday it returns nothing. To solve this problem I implemented a catch in the getToday function that checks for the day and if it is a weekend it returns the data from the previous Friday instead.